Crate subxt_signer

Subxt-signer

The main output from this crate is the sr25519::Keypair, which can be constructed from a bip39 phrase, secret URI or raw seed, and used to sign and verify arbitrary messages. This crate is aligned with how Substrate’s sp_core crate constructs and signs keypairs, but is lighter on dependencies and can support compilation to WASM with the web feature.

Enable the subxt feature to enable use of this sr25519::Keypair in signing subxt transactions for chains supporting sr25519 signatures.

Re-exports

pub use bip39;

Modules

ecdsa

An ecdsa keypair implementation.

eth

An ethereum keypair implementation.

sr25519

An sr25519 keypair implementation.

Structs

SecretUri

A secret uri (SURI) that can be used to generate a key pair.

Enums

DeriveJunction

A since derivation junction description. It is the single parameter used when creating a new secret key from an existing secret key and, in the case of SoftRaw and SoftIndex a new public key from an existing public key.

SecretUriError

This is returned if FromStr cannot parse a string into a SecretUri.

Constants

DEV_PHRASE

The root phrase for our publicly known keys.

Traits

ExposeSecret

Expose a reference to an inner secret

Type Aliases

SecretString

Secret string type.