subxt_signer

Module eth

An ethereum keypair implementation.

Modules

dev

Dev accounts, helpful for testing but not to be used in production, since the secret keys are known.

Structs

DerivationPath

A derivation path. This can be parsed from a valid derivation path string like "m/44'/60'/0'/0/0", or we can construct one using the helpers DerivationPath::empty() and DerivationPath::eth().

Keypair

An ethereum keypair implementation.

PublicKey

The public key for an Keypair key pair. This is the uncompressed variant of ecdsa::PublicKey.

Signature

A signature generated by Keypair::sign().

Enums

Error

An error handed back if creating a keypair fails.

Functions

verify

Verify that some signature for a message was created by the owner of the PublicKey.

Type Aliases

SecretKeyBytes

Bytes representing a private key.